+++ Decsys/ClientApp/src/stories/SurveyList/SortPanel.stories.js
@@ -1,13 +1,14 @@
 import React from "react";
-import { storiesOf } from "@storybook/react";
 import { action } from "@storybook/addon-actions";
 import { array } from "@storybook/addon-knobs";
-import SortPanel from "./SortPanel";
+import SortPanel from "components/SurveyList/SortPanel";
 
-storiesOf("Admin/SurveyList/SortPanel", module).add("Default", () => (
+export default { title: "Admin/SurveyList/SortPanel", component: SortPanel };
+
+export const Basic = () => (
   <SortPanel
     keys={array("Sort Field Keys", ["Name", "Order"])}
     state={{ key: "name" }}
     onSortButtonClick={action("Sort Button clicked")}
   />
-));
+);
